Handle cros test 2
